chore: restructure v3
