+++ packages/wallets/src/components/Base/InlineMessage/InlineMessage.scss
