+++ packages/wallets/src/components/Base/InlineMessage/index.ts
