+++ packages/wallets/src/features/cashier/screens/WalletDepositScreens/WalletDepositCryptoAddress/WalletDepositCryptoAddress.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect } from 'react';
 import QRCode from 'qrcode.react';
 import { useAuthorize, useDepositCryptoAddress } from '@deriv/api';
-import useDevice from '../../hooks/useDevice';
-import { Loader } from '../Loader';
-import WalletClipboard from '../WalletClipboard/WalletClipboard';
+import { Loader } from '../../../../../components/Loader';
+import { WalletClipboard } from '../../../../../components/WalletClipboard';
+import useDevice from '../../../../../hooks/useDevice';
 import './WalletDepositCryptoAddress.scss';
 
 const WalletDepositCryptoAddress = () => {
