+++ packages/wallets/src/features/cashier/screens/WalletTransactionsScreens/WalletTransactionsNoDataState/WalletTransactionsNoDataState.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useHistory } from 'react-router-dom';
-import NoRecentTransactions from '../../public/images/no-recent-transactions.svg';
+import NoRecentTransactions from '../../../../../public/images/no-recent-transactions.svg';
 import './WalletTransactionsNoDataState.scss';
 
 const WalletTransactionsNoDataState = () => {
