+++ packages/wallets/src/features/cashier/screens/WalletTransferScreens/WalletTransferFormInputField/WalletTransferFormInputField.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { useFormikContext } from 'formik';
-import useInputATMFormatter from '../../hooks/useInputATMFormatter';
+import useInputATMFormatter from '../../../../../hooks/useInputATMFormatter';
 import './WalletTransferFormInputField.scss';
 
 type TProps = {
